Adding IdentityManager and in memory implementation
