+++ OpenIDConnect.Host/packages.config
@@ -3,6 +3,7 @@
   <package id="Autofac" version="3.5.2" targetFramework="net452" />
   <package id="Autofac.WebApi2" version="3.4.0" targetFramework="net452" />
   <package id="AutoMapper" version="4.1.1" targetFramework="net452" />
+  <package id="IdentityManager" version="1.0.0-beta5-5" targetFramework="net452" />
   <package id="IdentityServer3" version="2.3.0" targetFramework="net452" />
   <package id="IdentityServer3.Admin" version="1.0.0-beta4" targetFramework="net452" />
   <package id="Microsoft.AspNet.WebApi.Client" version="5.2.3" targetFramework="net452" />
